fix style changing, small fixes

+++ src/media/scenario.h
@@ -1,6 +1,7 @@
 #ifndef SCENARIO_H
 #define SCENARIO_H
 
+#include <inttypes.h>
 #include <string>
 
 class Scenario {
@@ -11,7 +12,7 @@ public:
 	static void saveStack(const std::string &path);
 
 	static void execute(const std::string &loadPath);
-	static void jumpNext(const std::string &label, bool isCall, const std::string &fileName, size_t numLine);
+	static void jumpNext(const std::string &label, bool isCall, const std::string &fileName, uint32_t numLine);
 };
 
 #endif // SCENARIO_H
